pyproject.toml
[project]
name = "thermostat-proxy"
version = "1.0.0"
requires-python = ">=3.12"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest-homeassistant-custom-component>=0.13.195",
]
